HW7 added tests, fixed small bug in getting balance

+++ lessonseven/src/main/java/ru/petrovvich/study/model/ATM.java
@@ -14,5 +14,5 @@ public interface ATM {
 
     ATMResponse putCash(Integer sum);
 
-    Long getTotalBalanceInUnits();
+    Integer getTotalBalanceInUnits();
 }
